// src/app/models/Message.ts
/**
 * Структура информирующего сообщения для пользователя
 */
export interface Message {
  /**
   * Текст сообщения
   */
  messageText: string;

  /**
   * Предупреждающее ли сообщение -> для смены класса DOM объекта, содержащего сообщение
   */
  isWarning: boolean;
}