// src/app/models/ItemBookingData.ts
/**
 * Структура данных о бронировании товара
 */
export interface ItemBookingData {
  /**
   * Идентификатор
   */
  id: number;

  /**
   * Массив с датами на которые зарезервирован товар
   */
  dates: string[];
}